backend/requirements.txt
alembic==1.5.8
asyncpg==0.24.0
async-timeout==3.0.1
bcrypt==3.2.0
cachetools==4.2.1
cryptography==3.3.2
ecdsa==0.14.1
email-validator==1.1.2
fastapi==0.63.0
greenlet==1.0.0
h11==0.12.0
Mako==1.1.4
MarkupSafe==1.1.1
passlib==1.7.4
psycopg2-binary==2.8.6
pyasn1==0.4.8
pydantic==1.8.1
python-dateutil==2.8.1
python-jose==3.2.0
python-multipart==0.0.5
requests==2.25.1
rfc3986==1.4.0
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.4.5
SQLAlchemy-Utils==0.37.0
starlette==0.13.6
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
pytest==6.2.5